Add files via upload
Matrix Maestro is an advanced matrix calculator developed by Abdul Kabeer, a student of FAST National University. The application provides an intuitive interface to perform essential matrix operations such as:

Addition of matrices.
Subtraction of matrices.
Multiplication of matrices.
Calculation of the determinant.
Finding the transpose of a matrix.
Computing the inverse of a matrix (if it exists).
Matrix Maestro is designed to simplify complex mathematical computations, making it an invaluable tool for students, researchers, and professionals working with matrices in various fields like mathematics, computer science, and engineering.

With its focus on user-friendly design and robust functionality, Matrix Maestro aims to be a reliable companion for matrix-related tasks.